requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
